Custom GUI progress and code flexibility rework
Continuing work on adding the custom GUI to the script. My previous approach of Instantiating the Child GameObject was not working because VRChat does not give instantiated objects a network ID, so they can not perform any network functions. Therefore the Child must be present before run time. So there is now a button that can spawn a child if needed, and link it to the script. It checks if there is already a valid child so you cannot spawn more than one. It also notifies you if a child has not already been spawned.
The approach is to have a child gameobject that plays the sounds. This is because the network synchronisation methods provided by VRChat's API are Continuous, Manual and None. Continuous is great for when you do not really care about the timing of a network sync. However, Manual is much faster and responsive when sending small packets of information, which I am doing in this case. However, you cannot have a script with Manual synchronisation on a GameObject with the VRC Object Sync component on it. So the approach is to have a child object with manual sync play the sounds instead. This gets a much nicer and responsive outcome.