Custom UI values retained, Child script now handles audio
The values selected for 'play mode' and such are now correctly retained within the UI. Previously they would disappear when selecting another object and returning to the script, or when closing and reopening the program. These values are also now passed into the scripts for use.
This commit also contains the changes to the Audio playing sequence to make it so that the audio is played from the Child Object. The plan is to have the child object with manual sync method play the sounds as it is more responsive that way. The multiple sounds method was implemented and tested first, which is working as of this commit. Overall progress is good and not much is left to be done to get it completely working.