Massive changes to custom editor UI
I have had to change a lot of the code that I created thus far for the custom editor. This is because UdonSharp does not support the use of custom classes, so the classes I had created, and subsequently created a custom property drawer would cause a lot of errors when going in to play mode. But I have been looking for alternatives and have found the ReorderableList internal UI element. This allows me to customize how the list is shown without having to create any custom classes (even though I would prefer to :disappointed: ). But this commit contains the working code for the new ReorderableList. Currently it displays an Int array, with each int also displaying the corresponding AudioClip array value of the int index. So that pretty much is the Sequence tab working, now just need to finish the Random tab, which should be somewhat easy?