UI finished. Finishing code for handling tabs
As of this commit, the code for the UI is 99% complete (just need to change some variable names and change a button so that its label is not Test). This commit also contains the work on having the code work with the new tabs, such that when you select your options and run the game, it works. I was running into an issue with getting the Sequence mode working because the sequences were out of sync of someone joined late. I have a solution that is working for the most part but just needs a little bit more tuning. Notably, the solution works by having a flag (bool) that is set when the first serialization takes place. After this flag is set to true, each serialization should play the sound.