Support for Interact and Pickup objects
The final feature I wanted to add has now been added. Now, there is a dropdown menu at the top of the script that allows you to select which object the script is attached to. If the script is attached to a Interact, it will display additional options for adjusting the Proximity to interact with the object, and the Interaction Text that shows up when you hover over the object with a cursor. If the object type is a Pickup, it will not show these settings and disable the Interact event at runtime.